requirements.txt
google-generativeai==0.3.2
requests==2.31.0
colorama==0.4.6
rich==13.7.0
pyfiglet==1.0.2
gitpython==3.1.40
psutil==5.9.6
threading==1.0
json5==0.9.14
pyttsx3==2.90
speech_recognition==3.10.0
beautifulsoup4==4.12.2
selenium==4.15.2
nmap-python==1.5.4
scapy==2.5.0
cryptography==41.0.8
paramiko==3.3.1
python-whois==0.8.0
dnspython==2.4.2
shodan==1.30.1
censys==2.2.8